Abstract get client cert func
