+++ pkg/adsc/adsc.go
@@ -309,44 +309,11 @@ func getPrivateIPIfAvailable() net.IP {
 }
 
 func (a *ADSC) tlsConfig() (*tls.Config, error) {
-	var clientCert tls.Certificate
 	var clientCerts []tls.Certificate
 	var serverCABytes []byte
 	var err error
-	var certName string
 
-	var getClientCertificate func(requestInfo *tls.CertificateRequestInfo) (*tls.Certificate, error)
-
-	if a.cfg.SecretManager != nil {
-		getClientCertificate = func(requestInfo *tls.CertificateRequestInfo) (*tls.Certificate, error) {
-			tok, err := ioutil.ReadFile(a.cfg.JWTPath)
-			if err != nil {
-				log.Infof("Failed to get credential token in agent: %v", err)
-				tok = []byte("")
-			}
-
-			certName = fmt.Sprintf("(generated from %s)", a.cfg.JWTPath)
-			key, err := a.cfg.SecretManager.GenerateSecret(context.Background(), "agent",
-				cache.WorkloadKeyCertResourceName, string(tok))
-			if err != nil {
-				return nil, err
-			}
-			clientCert, err = tls.X509KeyPair(key.CertificateChain, key.PrivateKey)
-			if err != nil {
-				return nil, err
-			}
-			return &clientCert, nil
-		}
-	} else if a.cfg.CertDir != "" {
-		getClientCertificate = func(requestInfo *tls.CertificateRequestInfo) (*tls.Certificate, error) {
-			certName = a.cfg.CertDir + "/cert-chain.pem"
-			clientCert, err = tls.LoadX509KeyPair(certName, a.cfg.CertDir+"/key.pem")
-			if err != nil {
-				return nil, err
-			}
-			return &clientCert, nil
-		}
-	}
+	var getClientCertificate = getClientCertFn(a.cfg)
 
 	// Load the root CAs
 	if a.cfg.RootCert != nil {
@@ -372,18 +339,6 @@ func (a *ADSC) tlsConfig() (*tls.Config, error) {
 	serverCAs := x509.NewCertPool()
 	if ok := serverCAs.AppendCertsFromPEM(serverCABytes); !ok {
 		return nil, err
-	}
-
-	// If we supply an expired cert to the server it will just close the connection
-	// without useful message.  If the cert is obviously bogus, refuse to use it.
-	now := time.Now()
-	for _, cert := range clientCert.Certificate {
-		cert, err := x509.ParseCertificate(cert)
-		if err == nil {
-			if now.After(cert.NotAfter) {
-				return nil, fmt.Errorf("certificate %s expired %v", certName, cert.NotAfter)
-			}
-		}
 	}
 
 	shost, _, _ := net.SplitHostPort(a.url)
